Double precision floating point numbers everywhere

+++ src/Constants.java
@@ -2,13 +2,13 @@ public class Constants {
     /**
      * The speed of light in m/s
      */
-    public static final float SPEED_OF_LIGHT = 299_792_458f;
+    public static final double SPEED_OF_LIGHT = 299_792_458f;
 
     // List of refractive indexes
     // https://en.wikipedia.org/wiki/List_of_refractive_indices
-    public static final float REFRAC_VACUUM = 1f;
-    public static final float REFRAC_AIR = 1.000273f;
-    public static final float REFRAC_WATER = 1.333f;
-    public static final float REFRAC_DIAMOND = 2.417f;
-    public static final float REFRAC_GLASS = 1.52f;
+    public static final double REFRAC_VACUUM = 1f;
+    public static final double REFRAC_AIR = 1.000273f;
+    public static final double REFRAC_WATER = 1.333f;
+    public static final double REFRAC_DIAMOND = 2.417f;
+    public static final double REFRAC_GLASS = 1.52f;
 }
